pyproject.toml
# 项目元数据
[project]
name = "notebooks"
version = "0.1.0"
description = "基于 LangChain 和 LangGraph 的 AI 笔记本项目，用于构建和测试各种 AI 代理和工作流"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
requires-python = ">=3.12"

# 分类器，说明项目适用的领域、Python版本、许可证等[citation:1]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# 项目依赖
dependencies = [
    "debugpy>=1.8.17",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

# 开发依赖
[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

# UV 配置
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

# 项目入口
[project.scripts]
notebooks = "src.main:main"

# 代码风格配置
[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
strict = true

[tool.setuptools.packages.find]
where = ["src"]